requirements.txt
amqp==5.0.6
asgiref==3.3.4
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.1.0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-celery==3.1.17
django-cors-headers==3.7.0
django-jsonfield==1.4.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==20.9
prompt-toolkit==3.0.18
psycopg2-binary==2.9.1
PyJWT==2.3.0
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.7
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
